pyproject.toml
[tool.poetry]
name = "tbot"
version = "0.1.0"
description = ""
authors = ["Ivan Kovtun <jeanlacroix@ukr.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-telegram-bot = "^21.4"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.3.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
